requirements.txt
# requirements.txt

# Database connectors
pymysq
redis
pymongo
pymilvus

# Environment variable management
python-dotenv

# Additional useful packages
numpy
pandas
requests

# For potential future use with vector operations or machine learning
scikit-learn

# For handling JSON data
ujson

# For working with dates and times
python-dateutil

# For easier HTTP requests (if needed)
httpx

# For async programming (if needed)
aiohttp

# For data validation and settings management (recommended)
pydantic